file phone-sort.js moves to services some code refactoring in phones-page.js and catalogue.html
