+++ catalogue.html
@@ -33,7 +33,7 @@
         </section>
 
         <section>
-          <p>Shopping ShoppingCart</p>
+          <p>Shopping Cart</p>
           <ul>
             <li>Phone 1</li>
             <li>Phone 2</li>
